requirements.txt
# SoVAni Bot - Telegram бот для автоматизации работы с Wildberries и Ozon
# Список зависимостей Python

# Telegram Bot Framework
aiogram==2.25.1

# HTTP клиент для API запросов
aiohttp==3.8.5
requests==2.31.0
httpx==0.24.1

# OpenAI для генерации ответов
openai==0.27.8

# Планировщик задач
APScheduler==3.10.4

# Работа с Excel и данными
pandas==2.1.0
openpyxl==3.1.2

# База данных SQLite (встроена в Python, но для удобства)
# sqlite3 входит в стандартную библиотеку Python

# Работа с переменными окружения
python-dotenv==1.0.0

# Дополнительные утилиты
typing-extensions==4.7.1

# Логирование и дебаг (опционально)
colorlog==6.7.0

# Для работы с датами
python-dateutil==2.8.2

# Для улучшенной работы с JSON
ujson==5.8.0

# Дополнительная безопасность
cryptography==41.0.4

# Для работы с временными зонами
pytz==2023.3

# Инструменты разработки (опционально)
# black==23.7.0
# flake8==6.0.0
# pytest==7.4.0